# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Go modular monolith backend for emergency alert intake, ambulance dispatch, user management, and fleet readiness.",
        "title": "Dispatch Backend API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/blood/offers": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new blood offer for a requisition",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Blood"
                ],
                "summary": "Create blood offer",
                "parameters": [
                    {
                        "description": "Offer details",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dispatch_internal_modules_blood_application_dto.CreateBloodOfferRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/blood/requisitions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns paginated blood requisitions with search, sorting, and filters",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Blood"
                ],
                "summary": "List blood requisitions",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Page size",
                        "name": "page_size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search term",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "OPEN",
                            "BROADCASTING",
                            "MATCHED",
                            "PICKUP_ASSIGNED",
                            "COLLECTED",
                            "DELIVERED",
                            "CANCELLED",
                            "EXPIRED"
                        ],
                        "type": "string",
                        "description": "Filter by status",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by blood type",
                        "name": "blood_type",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "EMERGENCY",
                            "URGENT",
                            "ROUTINE"
                        ],
                        "type": "string",
                        "description": "Filter by urgency level",
                        "name": "urgency_level",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by created_at from (ISO 8601)",
                        "name": "filter[date_from]",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by created_at to (ISO 8601)",
                        "name": "filter[date_to]",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort by field: created_at, status, urgency_level",
                        "name": "sort_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Sort order: asc or desc",
                        "name": "sort_order",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a new blood requisition request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Blood"
                ],
                "summary": "Raise blood requisition",
                "parameters": [
                    {
                        "description": "Requisition details",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dispatch_internal_modules_blood_application_dto.CreateBloodRequisitionRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/blood/requisitions/{id}/offers": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the list of blood offers for a given requisition",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Blood"
                ],
                "summary": "List offers for requisition",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Blood Requisition ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/blood/requisitions/{id}/offers/{offerId}/accept": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Accepts a blood offer for a requisition and marks it as accepted",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Blood"
                ],
                "summary": "Accept blood offer",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Blood Requisition ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Blood Offer ID",
                        "name": "offerId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Optional audit actor",
                        "name": "payload",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/dispatch_internal_modules_blood_application_dto.AcceptOfferRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/pickup-assignments": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Assigns a pickup for an accepted blood offer, creating a new assignment record",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Blood"
                ],
                "summary": "Assign blood pickup",
                "parameters": [
                    {
                        "description": "Pickup assignment details",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dispatch_internal_modules_blood_application_dto.AssignBloodPickupRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/pickup-assignments/{assignmentId}/collect": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Marks a blood pickup assignment as collected by the assigned user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Blood"
                ],
                "summary": "Mark blood as collected",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pickup Assignment ID",
                        "name": "assignmentId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payload with requisition ID and optional actor user ID",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dispatch_internal_modules_blood_application_dto.MarkCollectedRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dispatch_internal_modules_blood_application_dto.AcceptOfferRequest": {
            "type": "object",
            "properties": {
                "actor_user_id": {
                    "type": "string"
                }
            }
        },
        "dispatch_internal_modules_blood_application_dto.AssignBloodPickupRequest": {
            "type": "object",
            "required": [
                "blood_requisition_id",
                "vehicle_type"
            ],
            "properties": {
                "ambulance_id": {
                    "type": "string"
                },
                "assigned_by_user_id": {
                    "type": "string"
                },
                "assigned_driver_user_id": {
                    "type": "string"
                },
                "blood_requisition_id": {
                    "type": "string"
                },
                "blood_requisition_offer_id": {
                    "type": "string"
                },
                "destination_facility_id": {
                    "type": "string"
                },
                "dispatch_assignment_id": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "pickup_site_id": {
                    "type": "string"
                },
                "vehicle_type": {
                    "type": "string",
                    "enum": [
                        "AMBULANCE",
                        "PICKUP",
                        "MOTORCYCLE",
                        "OTHER"
                    ]
                }
            }
        },
        "dispatch_internal_modules_blood_application_dto.CreateBloodOfferRequest": {
            "type": "object",
            "required": [
                "blood_group_code",
                "blood_product_code",
                "blood_requisition_id",
                "inventory_site_id",
                "units_offered"
            ],
            "properties": {
                "blood_group_code": {
                    "type": "string"
                },
                "blood_product_code": {
                    "type": "string"
                },
                "blood_requisition_id": {
                    "type": "string"
                },
                "contact_person_name": {
                    "type": "string"
                },
                "contact_phone": {
                    "type": "string"
                },
                "inventory_site_id": {
                    "type": "string"
                },
                "notes": {
                    "type": "string"
                },
                "offered_by_user_id": {
                    "type": "string"
                },
                "reserved_until": {
                    "type": "string"
                },
                "units_offered": {
                    "type": "integer",
                    "minimum": 1
                }
            }
        },
        "dispatch_internal_modules_blood_application_dto.CreateBloodRequisitionRequest": {
            "type": "object",
            "required": [
                "blood_group_code",
                "blood_product_code",
                "clinical_summary",
                "units_requested"
            ],
            "properties": {
                "blood_group_code": {
                    "type": "string"
                },
                "blood_product_code": {
                    "type": "string"
                },
                "clinical_summary": {
                    "type": "string"
                },
                "destination_facility_id": {
                    "type": "string"
                },
                "destination_lat": {
                    "type": "number"
                },
                "destination_lon": {
                    "type": "number"
                },
                "diagnosis": {
                    "type": "string"
                },
                "expires_at": {
                    "type": "string"
                },
                "incident_id": {
                    "type": "string"
                },
                "indication": {
                    "type": "string"
                },
                "parity_summary": {
                    "type": "string"
                },
                "patient_identifier": {
                    "type": "string"
                },
                "patient_name": {
                    "type": "string"
                },
                "reporter_phone": {
                    "type": "string"
                },
                "requested_by_user_id": {
                    "type": "string"
                },
                "requesting_facility_id": {
                    "type": "string"
                },
                "units_requested": {
                    "type": "integer",
                    "minimum": 1
                },
                "urgency_level": {
                    "type": "string",
                    "enum": [
                        "EMERGENCY",
                        "URGENT",
                        "ROUTINE"
                    ]
                }
            }
        },
        "dispatch_internal_modules_blood_application_dto.MarkCollectedRequest": {
            "type": "object",
            "required": [
                "blood_requisition_id"
            ],
            "properties": {
                "actor_user_id": {
                    "type": "string"
                },
                "blood_requisition_id": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}